Split Rect instance & creation funcs into 2 files
