+++ tests/unit/rect/test_rect_creation_helpers.py
@@ -3,7 +3,55 @@ from __future__ import annotations
 
 import pytest
 
-from arcade.types.rect import Kwargtangle
+from arcade.types.rect import LBWH, LRBT, XYWH, XYRR, Kwargtangle
+
+
+def test_make_LBWH():
+    r = LBWH(10, 10, 10, 10)
+    assert r.left == 10.0
+    assert r.bottom == 10.0
+    assert r.width == 10.0
+    assert r.height == 10.0
+    assert r.top == 20.0
+    assert r.right == 20.0
+    assert r.x == 15.0
+    assert r.y == 15.0
+
+
+def test_make_LRBT():
+    r = LRBT(10, 20, 10, 20)
+    assert r.left == 10.0
+    assert r.bottom == 10.0
+    assert r.width == 10.0
+    assert r.height == 10.0
+    assert r.top == 20.0
+    assert r.right == 20.0
+    assert r.x == 15.0
+    assert r.y == 15.0
+
+
+def test_make_XYWH():
+    r = XYWH(15, 15, 10, 10)
+    assert r.left == 10.0
+    assert r.bottom == 10.0
+    assert r.width == 10.0
+    assert r.height == 10.0
+    assert r.top == 20.0
+    assert r.right == 20.0
+    assert r.x == 15.0
+    assert r.y == 15.0
+
+
+def test_make_XYRR():
+    r = XYRR(15, 15, 5, 5)
+    assert r.left == 10.0
+    assert r.bottom == 10.0
+    assert r.width == 10.0
+    assert r.height == 10.0
+    assert r.top == 20.0
+    assert r.right == 20.0
+    assert r.x == 15.0
+    assert r.y == 15.0
 
 
 def test_kwargtangle_missing_args():
@@ -91,3 +139,4 @@ def test_kwargtangle_none_args():
 
     with pytest.raises(ValueError):
         _ = Kwargtangle(x=0, y=None, width=0, height=0)
+
